CodeProjectAI/requirements.txt
paho-mqtt
pillow
requests
pyyaml
